[Game] WIP: ClientFileManager - merge dev fixes
