+++ AAEmu.Game/IO/ClientSourceType.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+
+namespace AAEmu.Game.IO
+{
+    public enum ClientSourceType
+    {
+        Directory,
+        GamePak
+    }
+}
